move run and process_changes into the base class
